requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cssselect==1.3.0
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
readability-lxml==0.8.4.1
regex==2025.10.23
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
